refactor(store-core): extract cost-budget handlers into handlers/budget.ts (audit P2-3)
Fifth slice of the P2-3 split. Move the four cost-budget message handlers
(budget_warning / budget_exceeded / budget_resumed / budget_resume_ack)
into a new handlers/budget.ts.

Pure move, no logic change: the extracted block is byte-identical to the
removed lines (proven by diff). Deps are only ChatMessage (../types) and
nextMessageId (../utils) — no shared barrel helpers. The barrel re-exports
via `export * from './budget'`, so the public surface is unchanged.

handlers/index.ts: 5150 -> 5068 lines.
